pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2",
    "django-ninja>=1.4.1",
    "django-tailwind>=4.0.1",
    "gunicorn>=23.0.0",
    "huggingface-hub>=0.32.0",
    "ipykernel>=6.29.5",
    "keras>=3.9.2",
    "keras-hub>=0.20.0",
    "matplotlib>=3.10.1",
    "numpy>=2.1.3",
    "python-dotenv>=1.1.0",
    "tensorflow>=2.19.0",
    "whitenoise>=6.9.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
line-length = 100
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "migrations",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # flake8-simplify
    "SIM",
    # flake8-return
    "RET",
    # flake8-bugbear
    "B",
    # isort
    "I",
    # pyupgrade
    "UP",
]
ignore = [
    # line too long
    "E501",
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[dependency-groups]
dev = [
    "django-browser-reload>=1.18.0",
    "ruff>=0.11.10",
]
